requirements.txt
fastapi==0.115.8
mysql-connector-python==9.2.0
pandas==2.2.3
pydantic==2.10.6
requests==2.32.3
streamlit==1.42.0
uvicorn==0.34.0
pytest==8.3.4